menambah functions

+++ data.php
@@ -50,7 +50,6 @@ $konek = mysqli_connect("localhost", "root", "", "wpu-hut");
                 <th scope="col">Kategori</th>
                 <th scope="col">Deskripsi</th>
                 <th scope="col">Harga</th>
-                <th scope="col">Url Gambar</th>
                 <th scope="col">Action</th>
                 </tr>
             </thead>
@@ -66,9 +65,8 @@ $hasil = mysqli_query($konek, $sql);
                 <td><?=$row["kategori"];?></td>
                 <td><?=$row["deskripsi"];?></td>
                 <td><?=$row["harga"];?></td>
-                <td><?=$row["gambar"];?></td>
                 <td><a href='edit.php?kode=<?=$row["id"];?>' class="fa fa-edit" style="color:gold; text-decoration:none;"></a> |
-                <a href='hapus.php?id=<?=$row["id"];?>' class="fa fa-trash" style="color:tomato; text-decoration:none;"></a>
+                <a href="hapus.php?id=<?=$row["id"];?>" onclick="return confirm('yakin mau dihapus ?');" class="fa fa-trash" style="color:tomato; text-decoration:none;"></a>
                 </td>
                 </tr>
                 <?php }?>
